+++ packages/ragbits-core/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- move sources from ragbits-document-search to ragbits-core (#496)
+
 ## 0.13.0 (2025-04-02)
 - Make the score in VectorStoreResult consistent (always bigger is better)
 - Add router option to LiteLLMEmbedder (#440)